requirements.txt
git+https://github.com/dolfies/discord.py-self.git
art
aiohttp[speedups]
ujson
questionary
discord-py
requests